# patronazneSestre/REST/src/main/resources/doc/v1/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Aplikacija za patronazne sestre",
    "version" : "1.0.0",
    "title" : "Patronazne Sestre",
    "contact" : {
      "email" : "mo0429@student.uni-lj.si"
    }
  },
  "host" : "localhost",
  "basePath" : "/v1",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/kontakt/{id}" : {
      "get" : {
        "tags" : [ "Storitve za kotaktno osebo" ],
        "summary" : "Vrni kontaktno osebo",
        "description" : "Vraca specificno kontaktno osebo iz bazo",
        "operationId" : "returnKontakt",
        "consumes" : [ "application/xml", "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Kontakt"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Storitve za kotaktno osebo" ],
        "summary" : "Brisi kontaktno osebo",
        "description" : "Pobrise kontaktno osebo iz bazo",
        "operationId" : "deleteKontakt",
        "consumes" : [ "application/xml", "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/uporabnik/{id}" : {
      "get" : {
        "tags" : [ "Storitve za uporabnike" ],
        "summary" : "Vrni uporabnika",
        "description" : "Vrne uporabnika iz baze",
        "operationId" : "returnUporabnik",
        "consumes" : [ "application/xml", "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Uporabnik"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Storitve za uporabnike" ],
        "summary" : "Brisi uporabnika",
        "description" : "Brise uporabnika iz baze",
        "operationId" : "deleteUporabnik",
        "consumes" : [ "application/xml", "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Aktivnost" : {
      "type" : "object",
      "properties" : {
        "idaktivnost" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "vrstaObiska" : {
          "$ref" : "#/definitions/VrstaObiska"
        },
        "meritevs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Meritev"
          }
        }
      }
    },
    "Bolezen" : {
      "type" : "object",
      "properties" : {
        "idbolezen" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        }
      }
    },
    "DelovniNalog" : {
      "type" : "object",
      "properties" : {
        "iddelovniNalog" : {
          "type" : "integer",
          "format" : "int32"
        },
        "bolezen" : {
          "$ref" : "#/definitions/Bolezen"
        },
        "izvajalecZdravstvenihStoritev" : {
          "$ref" : "#/definitions/IzvajalecZdravstvenihStoritev"
        },
        "obisk" : {
          "$ref" : "#/definitions/Obisk"
        },
        "vrstaObiska" : {
          "$ref" : "#/definitions/VrstaObiska"
        },
        "zdravstveniDelavec" : {
          "$ref" : "#/definitions/ZdravstveniDelavec"
        },
        "materials" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Material"
          }
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        },
        "zdravilos" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Zdravilo"
          }
        }
      }
    },
    "IzvajalecZdravstvenihStoritev" : {
      "type" : "object",
      "properties" : {
        "idizvajalecZdravstvenihStoritev" : {
          "type" : "integer",
          "format" : "int32"
        },
        "hisnaStevilka" : {
          "type" : "string"
        },
        "naziv" : {
          "type" : "string"
        },
        "stevilkaIzvajalca" : {
          "type" : "string"
        },
        "ulica" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        },
        "posta" : {
          "$ref" : "#/definitions/Posta"
        },
        "zdravstveniDelavecs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ZdravstveniDelavec"
          }
        }
      }
    },
    "Kontakt" : {
      "type" : "object",
      "properties" : {
        "idkontakt" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ime" : {
          "type" : "string"
        },
        "priimek" : {
          "type" : "string"
        },
        "telefonskaStevilka" : {
          "type" : "string"
        },
        "sorodstvenoRazmerje" : {
          "$ref" : "#/definitions/SorodstvenoRazmerje"
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        }
      }
    },
    "Material" : {
      "type" : "object",
      "properties" : {
        "idmaterial" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        }
      }
    },
    "Meritev" : {
      "type" : "object",
      "properties" : {
        "idmeritev" : {
          "type" : "integer",
          "format" : "int32"
        },
        "merskaEnota" : {
          "type" : "string"
        },
        "obvezno" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "spodnjaMeja" : {
          "type" : "string"
        },
        "ugornjaMeja" : {
          "type" : "string"
        },
        "aktivnosts" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Aktivnost"
          }
        }
      }
    },
    "Obisk" : {
      "type" : "object",
      "properties" : {
        "idobisk" : {
          "type" : "integer",
          "format" : "int32"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        }
      }
    },
    "Okolis" : {
      "type" : "object",
      "properties" : {
        "idokolis" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "posta" : {
          "$ref" : "#/definitions/Posta"
        },
        "zdravstveniDelavecs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ZdravstveniDelavec"
          }
        }
      }
    },
    "Pacient" : {
      "type" : "object",
      "properties" : {
        "idpacient" : {
          "type" : "integer",
          "format" : "int32"
        },
        "hisnaStevilka" : {
          "type" : "string"
        },
        "ime" : {
          "type" : "string"
        },
        "priimek" : {
          "type" : "string"
        },
        "stevilkaZdravstvenegaZavarovanja" : {
          "type" : "string"
        },
        "ulica" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        },
        "kontakt" : {
          "$ref" : "#/definitions/Kontakt"
        },
        "pacient" : {
          "$ref" : "#/definitions/Pacient"
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        },
        "posta" : {
          "$ref" : "#/definitions/Posta"
        },
        "sorodstvenoRazmerje" : {
          "$ref" : "#/definitions/SorodstvenoRazmerje"
        },
        "spol" : {
          "$ref" : "#/definitions/Spol"
        },
        "uporabnik" : {
          "$ref" : "#/definitions/Uporabnik"
        }
      }
    },
    "Posta" : {
      "type" : "object",
      "properties" : {
        "idposta" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "izvajalecZdravstvenihStoritevs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/IzvajalecZdravstvenihStoritev"
          }
        },
        "okolis" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Okolis"
          }
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        }
      }
    },
    "SorodstvenoRazmerje" : {
      "type" : "object",
      "properties" : {
        "idsorodstvenoRazmerje" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "kontakts" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Kontakt"
          }
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        }
      }
    },
    "Spol" : {
      "type" : "object",
      "properties" : {
        "idspol" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        }
      }
    },
    "Uporabnik" : {
      "type" : "object",
      "properties" : {
        "iduporabnik" : {
          "type" : "integer",
          "format" : "int32"
        },
        "email" : {
          "type" : "string"
        },
        "geslo" : {
          "type" : "string"
        },
        "pacients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pacient"
          }
        },
        "vloga" : {
          "$ref" : "#/definitions/Vloga"
        },
        "zdravstveniDelavecs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ZdravstveniDelavec"
          }
        }
      }
    },
    "Vloga" : {
      "type" : "object",
      "properties" : {
        "idvloga" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "uporabniks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Uporabnik"
          }
        }
      }
    },
    "VrstaObiska" : {
      "type" : "object",
      "properties" : {
        "idvrstaObiska" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "aktivnosts" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Aktivnost"
          }
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        }
      }
    },
    "Zdravilo" : {
      "type" : "object",
      "properties" : {
        "idzdravilo" : {
          "type" : "integer",
          "format" : "int32"
        },
        "opis" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        }
      }
    },
    "ZdravstveniDelavec" : {
      "type" : "object",
      "properties" : {
        "idzdravstveniDelavec" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ime" : {
          "type" : "string"
        },
        "priimek" : {
          "type" : "string"
        },
        "delovniNalogs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DelovniNalog"
          }
        },
        "izvajalecZdravstvenihStoritev" : {
          "$ref" : "#/definitions/IzvajalecZdravstvenihStoritev"
        },
        "uporabnik" : {
          "$ref" : "#/definitions/Uporabnik"
        },
        "okolis" : {
          "$ref" : "#/definitions/Okolis"
        }
      }
    }
  }
}